refactor: removes providers and breakpoint indicator from root layout
